aider: Added the `glob` crate to the `Cargo.toml` file.
# Aider chat conversation:

USER: failed to resolve: use of undeclared crate or module `glob`

ASSISTANT: The error message "failed to resolve: use of undeclared crate or module `glob`" indicates that the `glob` crate is being used in the code, but it has not been declared in the `Cargo.toml` file.

To resolve this issue, you need to add the `glob` crate to your `Cargo.toml` file. However, as the `Cargo.toml` file is currently read-only, I cannot make this change for you.

Please add the `Cargo.toml` file to the chat so that I can provide you with the necessary *SEARCH/REPLACE* block to add the `glob` crate.

USER: I added these *read-write* files: Cargo.toml

ASSISTANT: To resolve the issue, we need to add the `glob` crate to your `Cargo.toml` file.

Here is the *SEARCH/REPLACE* block:

Cargo.toml
```toml
<<<<<<< SEARCH
[dependencies]
walkdir = "2.3.2"
git2 = "0.13.22"
=======
[dependencies]
walkdir = "2.3.2"
git2 = "0.13.22"
glob = "0.3.0"
>>>>>>> REPLACE
```

This block will add the `glob` crate to your dependencies. After applying this change, you should be able to use the `glob` crate in your code without any issues.